Compare IDG
Defence is sometimes listed as unfilled, sometimes as a three-person
council, and sometimes as Dbeibeh

Adel Mohamed Sultan is still listed as Minister on
https://csc.gov.ly/en/portfolio/ministry-of-marine-resources/

Ahmed Ali Muhammad Omar and Ahmed Abuhisa are both listed as Minister of
Industry throughout the period. I'm unsure if they're actually the same
person.

Bufarna resigned as Minister of Planning in March